Tidy up shady pie component
